Refactoring: separate elementary REST API endpoints from OLS-related endpoints
