+++ tests/integration/test_liveness_readiness.py
@@ -1,0 +1,32 @@
+"""Integration tests for /livenss and /readiness REST API endpoints."""
+
+import os
+from unittest.mock import patch
+
+import requests
+from fastapi.testclient import TestClient
+
+
+# we need to patch the config file path to point to the test
+# config file before we import anything from main.py
+@patch.dict(os.environ, {"OLS_CONFIG_FILE": "tests/config/valid_config.yaml"})
+def setup():
+    """Setups the test client."""
+    global client
+    from ols.app.main import app
+
+    client = TestClient(app)
+
+
+def test_liveness() -> None:
+    """Test handler for /liveness REST API endpoint."""
+    response = client.get("/liveness")
+    assert response.status_code == requests.codes.ok
+    assert response.json() == {"status": "1"}
+
+
+def test_readiness() -> None:
+    """Test handler for /readiness REST API endpoint."""
+    response = client.get("/readiness")
+    assert response.status_code == requests.codes.ok
+    assert response.json() == {"status": "1"}
